# Small optimizations on Scheduler creation