--- a/src/schedulers.jl
+++ b/src/schedulers.jl
@@ -85,18 +85,22 @@ struct ChunkingArgs{C, S <: Split}
     n::Int
     size::Int
     split::S
-    minsize::Union{Int, Nothing}
+    minsize::Int
 end
-ChunkingArgs(::Type{NoChunking}) = ChunkingArgs{NoChunking, NoSplit}(-1, -1, NoSplit(), nothing)
+ChunkingArgs(::Type{NoChunking}) = ChunkingArgs{NoChunking, NoSplit}(-1, -1, NoSplit(), -1)
 function ChunkingArgs(
         Sched::Type{<:Scheduler},
         n::MaybeInteger,
         size::MaybeInteger,
         split::Union{Symbol, Split};
-        minsize=nothing,
+        minsize::MaybeInteger,
         chunking
 )
     chunking || return ChunkingArgs(NoChunking)
+
+    if !isgiven(minsize)
+        minsize = -1
+    end
 
     if !isgiven(n) && !isgiven(size)
         n = default_nchunks(Sched)
@@ -126,7 +130,7 @@ chunking_mode(::ChunkingArgs{C}) where {C} = C
 has_n(ca::ChunkingArgs) = ca.n > 0
 has_size(ca::ChunkingArgs) = ca.size > 0
 has_split(::ChunkingArgs{C, S}) where {C, S} = S !== NoSplit
-has_minsize(ca::ChunkingArgs) = !isnothing(ca.minsize)
+has_minsize(ca::ChunkingArgs) = ca.minsize > 0
 chunking_enabled(ca::ChunkingArgs) = chunking_mode(ca) != NoChunking
 
 _chunkingstr(ca::ChunkingArgs{NoChunking}) = "none"
@@ -222,7 +226,7 @@ function DynamicScheduler(;
         chunksize::MaybeInteger = NotGiven(),
         chunking::Bool = true,
         split::Union{Split, Symbol} = Consecutive(),
-        minchunksize::Union{Nothing, Int}=nothing)
+        minchunksize::MaybeInteger = NotGiven())
     if isgiven(ntasks)
         if isgiven(nchunks)
             throw(ArgumentError("For the dynamic scheduler, nchunks and ntasks are aliases and only one may be provided"))
@@ -283,7 +287,7 @@ function StaticScheduler(;
         chunksize::MaybeInteger = NotGiven(),
         chunking::Bool = true,
         split::Union{Split, Symbol} = Consecutive(),
-        minchunksize::Union{Nothing, Int} = nothing)
+        minchunksize::MaybeInteger = NotGiven())
     if isgiven(ntasks)
         if isgiven(nchunks)
             throw(ArgumentError("For the static scheduler, nchunks and ntasks are aliases and only one may be provided"))
@@ -353,7 +357,7 @@ function GreedyScheduler(;
         chunksize::MaybeInteger = NotGiven(),
         chunking::Bool = false,
         split::Union{Split, Symbol} = RoundRobin(),
-        minchunksize::Union{Nothing, Int} = nothing)
+        minchunksize::MaybeInteger = NotGiven())
     if isgiven(nchunks) || isgiven(chunksize)
         chunking = true
     end