Now that we can completely skip the threading part when there is a single chunk, it would be good to get as close as possible to `mapreduce` (in performance) when we are in this case.

Base  `mapreduce`:
```julia
julia> @btime mapreduce(sin, +, $(1:10))
  65.331 ns (0 allocations: 0 bytes)
```

before:
```julia
julia> @btime tmapreduce(sin, +, $(1:10); minchunksize = 100)
  101.726 ns (3 allocations: 96 bytes)
```

With this PR:
```julia
julia> @btime tmapreduce(sin, +, $(1:10); minchunksize = 100)
  67.880 ns (0 allocations: 0 bytes)
```
This is done by changing the types to be "is bits" (in this case, i.e., not heap-allocated):

- `minsize` is negative instead of `nothing` in `ChunkingArgs`
- `threadpool` is a compile time symbol in `DynamicScheduler`. So `sched.threadpool` no longer works and is replaced by `threadpool(sched)`.

By doing some profiling, we can find that the small gap that remains with `mapreduce` is the call to `div` which allows to calculate the number of chunks. So it seems difficult to go further than that.


There is still an issue if we pass the scheduler as a symbol:
```julia
julia> @btime tmapreduce(sin, +, $(1:10); scheduler = :dynamic, minchunksize = 100)
  177.354 ns (5 allocations: 144 bytes)
```

This is because the symbol is a runtime variable. So to bind it to its associated type, there will necessarily be type instability. Sometimes Julia manages to do this by taking the symbol as a constant, but it seems rather random. There are several solutions:

```julia
julia> @btime tmapreduce(sin, +, $(1:10); scheduler = Val(:dynamic), minchunksize = 100)
  67.880 ns (0 allocations: 0 bytes)

julia> @btime tmapreduce(sin, +, $(1:10); scheduler = DynamicScheduler(minchunksize = 100))
  67.887 ns (0 allocations: 0 bytes)
```

One could go even further and implement the scheduler as a positional argument instead of a keyword one, i.e. `tmapreduce(sin, +, 1:10, DynamicScheduler(minchunksize = 100))`, but i am not really sure about that.