@arnaud-ma 2025-04-08
I just realized that simply inlining tmapreduce makes everything constant and there is almost nothing left to do